app/mini_project/backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Cors==4.0.1
PyJWT==2.9.0
psycopg2-binary==2.9.9
Werkzeug==3.0.3
